+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Added
 
-- add methods set_access_tokenW and set_phone_number_id for WhatsappClient
+- add methods set_access_token and set_phone_number_id for WhatsappClient
 
 ## [0.5.3] - 2024-03-13