+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Added
 
 - add methods set_access_token and set_phone_number_id for WhatsappClient
+- add methods request_code and verify_code for WhatsappClient
 
 ## [0.5.3] - 2024-03-13