+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 - add methods set_access_token and set_phone_number_id for WhatsappClient
 - add methods request_code and verify_code for WhatsappClient
 - add model struct VerificationRequest to help responding to webhook verification request
+- add getter and setter for WhatsappClient version
+- upgrade graph api version to v20.0
 
 ## [0.5.3] - 2024-03-13